--- a/Snappier.Benchmarks/UnalignedCopy128.cs
+++ b/Snappier.Benchmarks/UnalignedCopy128.cs
@@ -5,17 +5,18 @@ using Snappier.Internal;
 
 namespace Snappier.Benchmarks
 {
+
     [RyuJitX64Job]
     [InliningDiagnoser(false, new[] {"Snappier.Benchmarks"})]
-    public class UnalignedCopy64
+    public class UnalignedCopy128
     {
-        private readonly byte[] _buffer = new byte[16];
+        private readonly byte[] _buffer = new byte[32];
 
         [Benchmark]
-        public unsafe void Default()
+        public void Default()
         {
             ref byte ptr = ref _buffer[0];
-            CopyHelpers.UnalignedCopy64(ref ptr, ref Unsafe.Add(ref ptr, 8));
+            CopyHelpers.UnalignedCopy128(ref ptr, ref Unsafe.Add(ref ptr,16));
         }
     }
 }