--- a/Snappier.Benchmarks/UnalignedCopy64.cs
+++ b/Snappier.Benchmarks/UnalignedCopy64.cs
@@ -1,21 +1,22 @@
 ﻿using System.Runtime.CompilerServices;
 using BenchmarkDotNet.Attributes;
 using BenchmarkDotNet.Diagnostics.Windows.Configs;
+using BenchmarkDotNet.Jobs;
 using Snappier.Internal;
 
 namespace Snappier.Benchmarks
 {
     [RyuJitX64Job]
     [InliningDiagnoser(false, new[] {"Snappier.Benchmarks"})]
-    public class UnalignedCopy128
+    public class UnalignedCopy64
     {
-        private readonly byte[] _buffer = new byte[32];
+        private readonly byte[] _buffer = new byte[16];
 
         [Benchmark]
         public void Default()
         {
             ref byte ptr = ref _buffer[0];
-            CopyHelpers.UnalignedCopy128(ref ptr, ref Unsafe.Add(ref ptr,16));
+            CopyHelpers.UnalignedCopy64(ref ptr, ref Unsafe.Add(ref ptr, 8));
         }
     }
 }