Motivation
----------
There is room for improvement in how JIT handles these operations.

Modifications
-------------
Switch from using block copy operations to direct copy to/from a local temp variable of the correct size.

Results
-------

### IncrementalCopy 16 bytes with extra buffer room

BenchmarkDotNet=v0.13.4, OS=Windows 11 (10.0.22000.1455/21H2) Intel Core i7-10850H CPU 2.70GHz, 1 CPU, 12 logical and 6 physical cores .NET SDK=7.0.102
  [Host]                       : .NET 7.0.2 (7.0.222.60605), X64 RyuJIT AVX2
  MediumRun-.NET 6.0           : .NET 6.0.13 (6.0.1322.58009), X64 RyuJIT AVX2
  MediumRun-.NET 7.0           : .NET 7.0.2 (7.0.222.60605), X64 RyuJIT AVX2
  MediumRun-.NET Framework 4.8 : .NET Framework 4.8 (4.8.4515.0), X64 RyuJIT VectorSize=256

IterationCount=15  LaunchCount=2  WarmupCount=10

| Method |                          Job |            Runtime |      Mean |     Error |    StdDev | Ratio | RatioSD | Code Size |
|------- |----------------------------- |------------------- |----------:|----------:|----------:|------:|--------:|----------:|
|    Old |           MediumRun-.NET 6.0 |           .NET 6.0 |  3.251 ns | 0.0714 ns | 0.1023 ns |  1.00 |    0.00 |     689 B |
|    New |           MediumRun-.NET 6.0 |           .NET 6.0 |  2.955 ns | 0.0159 ns | 0.0239 ns |  0.91 |    0.03 |     496 B |
|        |                              |                    |           |           |           |       |         |           |
|    Old |           MediumRun-.NET 7.0 |           .NET 7.0 |  3.097 ns | 0.1127 ns | 0.1686 ns |  1.00 |    0.00 |     572 B |
|    New |           MediumRun-.NET 7.0 |           .NET 7.0 |  2.981 ns | 0.1097 ns | 0.1573 ns |  0.97 |    0.03 |     423 B |
|        |                              |                    |           |           |           |       |         |           |
|    Old | MediumRun-.NET Framework 4.8 | .NET Framework 4.8 | 16.226 ns | 0.0271 ns | 0.0388 ns |  1.00 |    0.00 |     657 B |
|    New | MediumRun-.NET Framework 4.8 | .NET Framework 4.8 |  8.235 ns | 0.0114 ns | 0.0163 ns |  0.51 |    0.00 |     390 B |

### Block decompression of Alice

BenchmarkDotNet=v0.13.4, OS=Windows 11 (10.0.22000.1455/21H2) Intel Core i7-10850H CPU 2.70GHz, 1 CPU, 12 logical and 6 physical cores .NET SDK=7.0.102
  [Host]                       : .NET 7.0.2 (7.0.222.60605), X64 RyuJIT AVX2
  MediumRun-.NET 6.0           : .NET 6.0.13 (6.0.1322.58009), X64 RyuJIT AVX2
  MediumRun-.NET 7.0           : .NET 7.0.2 (7.0.222.60605), X64 RyuJIT AVX2
  MediumRun-.NET Framework 4.8 : .NET Framework 4.8 (4.8.4515.0), X64 RyuJIT VectorSize=256

IterationCount=15  LaunchCount=2  WarmupCount=10

| Method |                          Job |            Runtime |      Mean |    Error |   StdDev |    Median | Ratio | RatioSD | Code Size |
|------- |----------------------------- |------------------- |----------:|---------:|---------:|----------:|------:|--------:|----------:|
|    Old |           MediumRun-.NET 6.0 |           .NET 6.0 | 109.37 us | 1.141 us | 1.636 us | 108.88 us |  1.00 |    0.00 |     805 B |
|    New |           MediumRun-.NET 6.0 |           .NET 6.0 | 102.41 us | 1.243 us | 1.861 us | 103.44 us |  0.94 |    0.02 |     805 B |
|        |                              |                    |           |          |          |           |       |         |           |
|    Old |           MediumRun-.NET 7.0 |           .NET 7.0 | 105.09 us | 1.413 us | 2.071 us | 105.04 us |  1.00 |    0.00 |     780 B |
|    New |           MediumRun-.NET 7.0 |           .NET 7.0 |  96.04 us | 1.181 us | 1.731 us |  96.32 us |  0.91 |    0.02 |     780 B |
|        |                              |                    |           |          |          |           |       |         |           |
|    Old | MediumRun-.NET Framework 4.8 | .NET Framework 4.8 | 115.33 us | 0.934 us | 1.368 us | 114.85 us |  1.00 |    0.00 |   1,376 B |
|    New | MediumRun-.NET Framework 4.8 | .NET Framework 4.8 | 109.05 us | 1.166 us | 1.634 us | 108.31 us |  0.95 |    0.02 |   1,376 B |